lambda-style/requirements.txt
Pillow
chainer==1.23
numpy
